reworked selection architecture
instead of manually coding which widget should be selected when you move from a widget,
there is now an invisible cursor that moves around the screen. so when youre on one widget and
move to the right, the cursor will move right until it hits one of the widgets and then stops.
this accidentally enabled the possibility to use your mouse to click on widgets, so that's neat.